Home

Contents

About NMR.jl

NMR.jl

Package Features

Manual Outline

Feedback

Citing NMR.jl

User Manual

Getting started

Loading NMR Data

Phase Correction

Cutting Regions

Roadmap to Version 1.0

Home

Contents

Example documentation

Example documentation

NMR — Module

module NMR

provides routines for processing 1D NMR spectra, as well as for simple Hilbert space spin dynamics simulations.

Version History

v0.9.0

added JEOL data reading

v0.8.0

added GISSMO library interface

v0.7.1

added readBrukerParameterFile()

v0.6.5

added medianBaseline() for baseline correction

added an entry for DSS in the HMDB table

NMR.FourierTransform — Function

FourierTransform(d::Data1D)

performs an FFT assuming that the input data set d is a free induction decay, and returns a DataSet object with the resulting complex data.