Reference

NMR.AutoPhaseCorrect — Method

AutoPhaseCorrect(spect::Data1D; exclude=false,threshold=10)

performs zero- and first-order phase correction of spect using a the algorithm by van der Waals and Geerens, Journal of Magnetic Resonance 86 (1990) 127-154. It works by recognising the peaks in the magnitude mode spectrum and then representing the phase at each peak location through a linear regression.

Passing a pair of indices to exclude disregards any peaks between these two boundaries. This is useful in spectra that contain a solvent artefact that may not be phased correctly.

threshold defines a threshold for peak detection (cf. NMR.peaks()).

NMR.AutoPhaseCorrectChen — Method

AutoPhaseCorrectChen(spect::Data1D;verbose=false) performs zero- and first-order phase correction of spect using a minimum entropy algorithm by Chen et al. in Journal of Magnetic Resonance 158 (2002) 164–168. It uses the Nelder-Mead algorithm, as implemented in the Optim.jl package.

NMR.BaseLineCorrect — Method

Automatic Baseline Correction

BaseLineCorrect(spect::Data1D;regions=128,kfactor=5,wdw=32) corrects the base line of spect using the algorithm by S. Golotvin, A. Williams, J. Magn. Reson. 146 (2000) 122-125.

Optional Parameters:

regions: the number of regions the spectrum is divided into to estimate the noise variance. This should be chosen large enough so that at least one of these regions is free from signals, and shows a flat baseline. The variance $\\sigma^2_{min}$ in this region is then used to distinguish signals and noise.

wdw: the window size around each data point that is used to decide whether the data point is part of a signal, or should be counted as base line.

kfactor: a point will be counted as part of the baseline if and only if the local variance in its window is less than kfactor*$\\sigma^2_{min}$.

output: defaults to :spectrum. output=:points will return an array of points, which indicate baseline positions (locations free from peaks) as determined by the algorithm. This is useful in situations where a large number of similar spectra need to be processed.

Notes

if the input data is complex, the imaginary part is ignored, and the result is always real.

the first and last point of spect are always used as part of the baseline. The first and last points of the output are therefore guaranteed to be zero. This helps to keep the interpolation numerically stable.

Baseline Correction With Known Locations

BaseLineCorrect(spect::Data1D,pts::Array{Float64,1}) uses the points in pts to determine the baseline. This should be a list of positions in the spectrum that are known not to contain signal; it can be obtained by using the option output=:points in the first method of BaseLineCorrect().

NMR.FourierTransform — Method

FourierTransform(d::Data1D)

performs an FFT assuming that the input data set d is a free induction decay, and returns a DataSet object with the resulting complex data.

NMR.FromXML — Method

 FromXML(e::XMLElement)

converts XML representation to PeakStruct

NMR.PeakSpect — Method

function PeakSpect(p,i,r;lw=0.0001)

compute the the spectrum with Lorentzian peaks at locations given in array p and (complex) amplitudes in array i, over the range r, which can be either an array or a range. lwis the line width. AData1D` object is returned.

NMR.PhaseCorrect — Method

PhaseCorrect(spect::Data1D;Ph0=0.0,Ph1=0.0,Pivot=0.0) returns a phase corrected DataSet, using the values indicated. The parameters are given in rad, k, and rad/k, where k are the units of the spectral domain (typically, Hz or ppm).

NMR.Spectrum — Method

    function Spectrum(ρ,H,Ψ;tol=1e-3)

compute the frequencies and intensities of the spectrum given by the the initial density operator ρ, the Hamiltonian H, and the observation operator Ψ. The results are returned as a tuple (freq,int) of two one-dimensional arrays. tol is a cutoff; transitions with absolute amplitudes less than this value are suppressed.

NMR.ToXML — Method

ToXML(p)

generates an XML structure containing the peak intensities, positions, and widths contained in the PeakStruct p.

NMR.clorentzian — Method

clorentzian(x0::Float64,σ::Float64,x::Float64)

returns the complex Lorentzian

$L(x) = \sqrt{\sigma}\dfrac{1+i\sqrt{\sigma}(x-x_0)}{1+\sigma(x-x_0)^2}$

NMR.derivative — Method

derivative(spect::Data1D) computes the derivative of spect and returns it as another Data1D object

NMR.entropy — Method

entropy(spect::Data1D) computes the entropy of the first derivative in the real part of an NMR spectrum as defined by Chen et al. in Journal of Magnetic Resonance 158 (2002) 164–168. This quantity can be optimised with respect to zero- and first-order phase correction for automatic (unsupervised) phase correction.

NMR.hard_shift — Method

function hard_shift(d::Data1D,δ::Number)

shifts the data in d by δ. The data is re-interpolated as needed. The resulting Data1D object is guaranteed to have the same index as d, such that they are compatible for the purposes of addition etc. Data shifted out of the window is lost; data points shifted into view are set to zero.

NMR.integral — Method

integral(spect::Data1D) computes the numerical integral of spect by the trapezoid rule.

NMR.medianBaseline — Method

function medianBaseline(s::NMR.Data1D;wdw=length(s)>>6)

Compute baseline for the real part of s by the algorithm of M. S. Friedrichs, Journal of Biomolecular NMR, 5 (1995) 147 153. The window is automatically calculated as $\Delta\omega/2^6$, unless overridden by giving a value to the optional key wdw.

NMR.peaks — Method

peaks(dinput::Data1D;threshold=1,athresh=1,regions=128)

identifies the peaks in dinput. Peak positions are located roughly by searching for the downward zero crossings of the first derivative. The data points surrounding each peak are then used for a quadratic regression, from which the position, curvature, and height of the maximum are interpolated. The signal dinput is then decomposed into a linear superposition of Lorentzian peaks at the determined positions and with the determined curvatures.

peaks() returns a PeakStruct data structure

where positions and heights are the interpolated peak positions and heights, respectively; widths are the peak curvatures, which correspond to 1/FWHM line widths of the Lorentzian peaks. intensities are the peak intensities resulting from deconvolution. These values are proportional to the underlying concentrations in an NMR spectrum. Finally, deconvolution contains the deconvoluted approximation to the original input signal.

NMR.readBrukerFID — Method

Bruker.readBrukerFID(s::String; format=Float64) reads a Bruker fid or ser file and returns the data contained in it as a complex array. For files generated by TopSpin 4.0 and higher, the default format (Float64) will work. For TopSpin 2.0, use format=Int32.

NMR.readBrukerParameterFile — Method

readBrukerParameterFile(s::String)

reads a Bruker JCAMP-DX parameter file, and returns a dict with the values. Arrays such as D0 D1 D2 etc are returned as a Julia array. Note that the Bruker parameter D0 appears as param["D"][1] due to Julia using 1-based arrays.

NMR.shift — Method

shift(p::PeakStruct,delta::Real)

Shift the positions of all peaks in p by delta

NMR.smooth — Function

smooth(d::Data1D,n=10)

returns a smoothed version of the input data set d, using convolution with a Gaussian peak.

NMR.Data1D — Type

Data1D(f<:AbstractArray,start<:Real,stop<:Real)

returns a data structure to hold a 1D NMR dataset either in the time or the frequency domain. Data1D objects store an index (range of time points or frequency points) along with the corresponding y-axis data. Data1D objects can be added and subtracted from one another, as long as their indices match.

NMR.PeakStruct — Type

mutable struct PeakStruct

Data structure to hold a peak list.

NMR.GISSMO.Hamiltonian — Method

function Hamiltonian(fn::String;freq=600.0,ctr=4.8)
function Hamiltonian(system::XMLElement;freq=600.0,ctr=4.8)

computes the Hamiltonian of an entry in the GISSMO database and returns it in the natural basis as a sparse matrix. fn is the GISSMO reference name of the compound. Keyword parameters are used to indicate spectrometer base frequency (in MHz) and the spectral zero point (carrier position) in ppm. maxSpin represents the maximum size of the spin system that will be simulated. If the spin system size exceeds maxSpin, the function returns (-1,nothing).

NMR.HMDB.decompositionMatrix — Method

decompositionMatrix(names,range)

compute a decomposition matrix B such that B s is the vector of concentrations of the metabolites listed in names. The pseudo-inverse is used for this purpose.

NMR.HMDB.matchPeaks — Method

(score,iscore,alpha,std) = matchPeaks(p::NMR.PeakStruct,ref::HMDBpeaks;tol=0.001)

match the peaks in p to the reference ref. A tuple of four values is returned:

score: sum of the intensities of the peaks found, normalised by the height of the largest reference peak. Missing peaks are counted as negative

iscore: score normalised by the total intensity of the reference peaks (range -1...1)

alpha: fitted normalised concentration

std: mean error of fitted normalised concentration

NMR.HMDB.matchReport — Method

xml=matchReport(p::NMR.PeakStruct,tol=0.01,AutoShift=true,iscoreCutoff=0,Id="",verbose=true)

match peaks in p against all metabolites in HMDB.refPeaks, and create a table of the results. If AutoShift is set to true (default), the p is automatically shfited for the greatest overlap with D-Glucose. tol is the matching tolerance for peak identification. Metabolites with iscores below iscoreCutoff are not reported. Metabolites are listed in decreasing order of concentration. matchReport() returns an XML object containing the listed data. Id is incorporated into the XML object as an attribute, to facilitate later identification of the match report.

NMR.HMDB.refSpectrum — Method

s::Data1D = refSpectrum(name::String,range;lw=0.01)

compute a reference spectrum over range using the peak positions in reference spectrum name. A Lorentzian line with half width lw is used. The reference spectrum is returned as a real Data1D object, unless the option complex=true is given.